Refactor network modification tests (#167)
