+++ src/main/java/org/gridsuite/modification/server/entities/equipment/modification/attribute/BooleanEquipmentAttributeModificationEntity.java
@@ -6,14 +6,14 @@
  */
 package org.gridsuite.modification.server.entities.equipment.modification.attribute;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.gridsuite.modification.server.dto.EquipmentAttributeModificationInfos;
+
 import javax.persistence.Entity;
 import javax.persistence.ForeignKey;
 import javax.persistence.PrimaryKeyJoinColumn;
 import javax.persistence.Table;
-
-import com.powsybl.iidm.network.IdentifiableType;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 /**
  * @author Slimane Amar <slimane.amar at rte-france.com>
@@ -24,7 +24,11 @@ import lombok.NoArgsConstructor;
 @Table(name = "booleanEquipmentAttributeModification")
 @PrimaryKeyJoinColumn(foreignKey = @ForeignKey(name = "boolean_equipment_attribute_modification_id_fk_constraint"))
 public class BooleanEquipmentAttributeModificationEntity extends EquipmentAttributeModificationEntity<Boolean> {
-    public BooleanEquipmentAttributeModificationEntity(String equipmentId, String attributeName, Boolean attributeValue, IdentifiableType equipmentType) {
-        super(equipmentId, attributeName, attributeValue, equipmentType);
+
+    public BooleanEquipmentAttributeModificationEntity(EquipmentAttributeModificationInfos equipmentAttributeModificationInfos) {
+        super(equipmentAttributeModificationInfos.getEquipmentId(),
+            equipmentAttributeModificationInfos.getEquipmentAttributeName(),
+            (boolean) equipmentAttributeModificationInfos.getEquipmentAttributeValue(),
+            equipmentAttributeModificationInfos.getEquipmentType());
     }
 }
