+++ src/main/java/org/gridsuite/modification/server/entities/equipment/modification/attribute/StringEquipmentAttributeModificationEntity.java
@@ -6,14 +6,14 @@
  */
 package org.gridsuite.modification.server.entities.equipment.modification.attribute;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.gridsuite.modification.server.dto.EquipmentAttributeModificationInfos;
+
 import javax.persistence.Entity;
 import javax.persistence.ForeignKey;
 import javax.persistence.PrimaryKeyJoinColumn;
 import javax.persistence.Table;
-
-import com.powsybl.iidm.network.IdentifiableType;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 /**
  * @author Slimane Amar <slimane.amar at rte-france.com>
@@ -24,7 +24,13 @@ import lombok.NoArgsConstructor;
 @Table(name = "stringEquipmentAttributeModification")
 @PrimaryKeyJoinColumn(foreignKey = @ForeignKey(name = "string_equipment_attribute_modification_id_fk_constraint"))
 public class StringEquipmentAttributeModificationEntity extends EquipmentAttributeModificationEntity<String> {
-    public StringEquipmentAttributeModificationEntity(String equipmentId, String attributeName, String attributeValue, IdentifiableType equipmentType) {
-        super(equipmentId, attributeName, attributeValue, equipmentType);
+
+    public StringEquipmentAttributeModificationEntity(EquipmentAttributeModificationInfos equipmentAttributeModificationInfos) {
+        super(equipmentAttributeModificationInfos.getEquipmentId(),
+            equipmentAttributeModificationInfos.getEquipmentAttributeName(),
+            equipmentAttributeModificationInfos.getEquipmentAttributeValue() == null ?
+                (String) equipmentAttributeModificationInfos.getEquipmentAttributeValue() :
+                equipmentAttributeModificationInfos.getEquipmentAttributeValue().toString(),
+            equipmentAttributeModificationInfos.getEquipmentType());
     }
 }
